fixes for changed LLVM API

+++ lib/bitcode/native_device_binding.cpp
@@ -1,6 +1,6 @@
 #include "pacxx/detail/device/DeviceCode.h"
 #include "pacxx/detail/device/DeviceFunctionDecls.h"
-#include <cmath>
+#include <math.h>
 
 __forceinline__ __index_t get_global_id(unsigned int dimindx) {
   switch (dimindx) {
@@ -68,9 +68,9 @@ __forceinline__ __index_t get_grid_size(unsigned int dimindx) {
 __forceinline__ void barrier(unsigned int) { __pacxx_barrier(); }
 
 extern "C" __forceinline__ double rsqrt(double val) {
-  return 1.0 / std::sqrt(val);
+  return 1.0 / sqrt(val);
 }
 
 extern "C" __forceinline__ float rsqrtf(float val) {
-  return 1.0f / std::sqrt(val);
+  return 1.0f / sqrtf(val);
 }
